% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{Base}
\alias{Base}
\title{\code{Base} class}
\description{
the \code{Base} class used by other objects; useful
for extensions of the package, not for basic
interactions with the package
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Base-new}{\code{Base$new()}}
\item \href{#method-Base-get_name}{\code{Base$get_name()}}
\item \href{#method-Base-get_type}{\code{Base$get_type()}}
\item \href{#method-Base-get_model}{\code{Base$get_model()}}
\item \href{#method-Base-set_model}{\code{Base$set_model()}}
\item \href{#method-Base-get_method}{\code{Base$get_method()}}
\item \href{#method-Base-set_method}{\code{Base$set_method()}}
\item \href{#method-Base-get_pi_method}{\code{Base$get_pi_method()}}
\item \href{#method-Base-set_pi_method}{\code{Base$set_pi_method()}}
\item \href{#method-Base-get_level}{\code{Base$get_level()}}
\item \href{#method-Base-set_level}{\code{Base$set_level()}}
\item \href{#method-Base-get_B}{\code{Base$get_B()}}
\item \href{#method-Base-set_B}{\code{Base$set_B()}}
\item \href{#method-Base-get_nb_hidden}{\code{Base$get_nb_hidden()}}
\item \href{#method-Base-set_nb_hidden}{\code{Base$set_nb_hidden()}}
\item \href{#method-Base-get_nodes_sim}{\code{Base$get_nodes_sim()}}
\item \href{#method-Base-set_nodes_sim}{\code{Base$set_nodes_sim()}}
\item \href{#method-Base-get_activ}{\code{Base$get_activ()}}
\item \href{#method-Base-set_activ}{\code{Base$set_activ()}}
\item \href{#method-Base-set_engine}{\code{Base$set_engine()}}
\item \href{#method-Base-get_engine}{\code{Base$get_engine()}}
\item \href{#method-Base-get_params}{\code{Base$get_params()}}
\item \href{#method-Base-get_seed}{\code{Base$get_seed()}}
\item \href{#method-Base-set_seed}{\code{Base$set_seed()}}
\item \href{#method-Base-summary}{\code{Base$summary()}}
\item \href{#method-Base-clone}{\code{Base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-new"></a>}}
\if{latex}{\out{\hypertarget{method-Base-new}{}}}
\subsection{Method \code{new()}}{
Create a new object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$new(
  name = "Base",
  type = "none",
  model = NULL,
  method = NULL,
  X_train = NULL,
  y_train = NULL,
  pi_method = c("none", "splitconformal", "kdesplitconformal", "bootsplitconformal",
    "jackknifeplus", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  level = 95,
  B = 100,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  seed = 123
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the class}

\item{\code{type}}{type of supervised learning method implemented}

\item{\code{model}}{fitted model}

\item{\code{method}}{supevised learning method}

\item{\code{X_train}}{training set features}

\item{\code{y_train}}{training set response}

\item{\code{pi_method}}{type of prediction interval in c("splitconformal",
"kdesplitconformal", "bootsplitconformal", "jackknifeplus",
"kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
"surrjackknifeplus")}

\item{\code{level}}{an integer; the level of confidence}

\item{\code{B}}{an integer; the number of simulations when \code{level} is not \code{NULL}}

\item{\code{nb_hidden}}{number of nodes in the hidden layer, for construction of a quasi-
randomized network}

\item{\code{nodes_sim}}{type of 'simulations' for hidden nodes, if \code{nb_hidden} > 0;
takes values in c("sobol", "halton", "unif")}

\item{\code{activ}}{activation function's name for the hidden layer, in the construction
of a quasi-randomized network; takes values in c("relu", "sigmoid", "tanh", "
leakyrelu", "elu", "linear")}

\item{\code{engine}}{contains fit and predict lower-level methods for the given \code{method};
do not modify by hand}

\item{\code{params}}{additional parameters passed to \code{method} when calling \code{fit}}

\item{\code{seed}}{an integer; reproducibility seed for methods that include
randomization}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Base} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_type}{}}}
\subsection{Method \code{get_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_model}{}}}
\subsection{Method \code{set_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_model(model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_method"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_method}{}}}
\subsection{Method \code{get_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_method()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_method"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_method}{}}}
\subsection{Method \code{set_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_method(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_pi_method"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_pi_method}{}}}
\subsection{Method \code{get_pi_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_pi_method()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_pi_method"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_pi_method}{}}}
\subsection{Method \code{set_pi_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_pi_method(pi_method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_level"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_level}{}}}
\subsection{Method \code{get_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_level()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_level"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_level}{}}}
\subsection{Method \code{set_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_level(level)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_B"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_B}{}}}
\subsection{Method \code{get_B()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_B()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_B"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_B}{}}}
\subsection{Method \code{set_B()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_B(B)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_nb_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_nb_hidden}{}}}
\subsection{Method \code{get_nb_hidden()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_nb_hidden()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_nb_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_nb_hidden}{}}}
\subsection{Method \code{set_nb_hidden()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_nb_hidden(nb_hidden)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_nodes_sim"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_nodes_sim}{}}}
\subsection{Method \code{get_nodes_sim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_nodes_sim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_nodes_sim"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_nodes_sim}{}}}
\subsection{Method \code{set_nodes_sim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_nodes_sim(nodes_sim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_activ"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_activ}{}}}
\subsection{Method \code{get_activ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_activ()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_activ"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_activ}{}}}
\subsection{Method \code{set_activ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_activ(activ)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_engine"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_engine}{}}}
\subsection{Method \code{set_engine()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_engine(engine)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_engine"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_engine}{}}}
\subsection{Method \code{get_engine()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_engine()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_params}{}}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-get_seed"></a>}}
\if{latex}{\out{\hypertarget{method-Base-get_seed}{}}}
\subsection{Method \code{get_seed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$get_seed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-set_seed"></a>}}
\if{latex}{\out{\hypertarget{method-Base-set_seed}{}}}
\subsection{Method \code{set_seed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$set_seed(seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Base-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$summary(
  X,
  show_progress = TRUE,
  class_name = NULL,
  class_index = NULL,
  y = NULL,
  type_ci = c("student", "nonparametric", "bootstrap", "conformal"),
  cl = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Base-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
