% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern32.R
\name{fit_matern32_classification}
\alias{fit_matern32_classification}
\title{X <- as.matrix(iris[,-ncol(iris)])
y <- iris$Species}
\usage{
fit_matern32_classification(
  x,
  y,
  lambda = 0.1,
  l = NULL,
  method = c("chol", "solve", "svd", "eigen"),
  with_kmeans = FALSE,
  centers = NULL,
  centering = FALSE,
  seed = 123,
  cl = NULL,
  ...
)
}
\description{
set.seed(123)
(index_train <- base::sample.int(n = nrow(X), size = floor(0.7*nrow(X)), replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]
dim(X_train)
dim(X_test)
}
\details{
fit_obj <- fit_matern32_classification(x = X_train, y = y_train, lambda = 0.1)

preds <- predict_matern32(fit_obj, X_test)
}
