% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern32.R
\name{fit_matern32_regression}
\alias{fit_matern32_regression}
\title{library(MASS)}
\usage{
fit_matern32_regression(
  x,
  y,
  lambda = 10^seq(-10, 10, length.out = 100),
  l = NULL,
  method = c("chol", "solve", "svd", "eigen"),
  with_kmeans = FALSE,
  centers = NULL,
  centering = FALSE,
  seed = 123,
  cl = NULL,
  ...
)
}
\description{
X <- as.matrix(Boston[,-ncol(Boston)])
y <- Boston$medv
}
\details{
set.seed(123)
(index_train <- base::sample.int(n = nrow(X), size = floor(0.7*nrow(X)), replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]
dim(X_train)
dim(X_test)

(lams <- 10^seq(-10, 10, length.out = 50))

fit_obj <- fit_matern32_regression(x = X_train, y = y_train, lambda = lams)

par(mfrow=c(1, 2))
plot(log(fit_obj$lambda), fit_obj$GCV, type = 'l', main = "GCV", ylab = "GCV")
matplot(log(fit_obj$lambda), t(fit_obj$coef), type = 'l', main = "coefficients = f(lambda)", xlab = "log(lambda)", ylab = "coefs")
abline(h = 0, lty = 2, lwd = 2, col = "red")

preds <- learningmachine::predict_matern32(fit_obj = fit_obj, newx = X_test)
RMSEs <- sqrt(colMeans((preds - y_test)^2))
plot(log(lams), log(RMSEs), type = 'l')
print(lams[which.min(log(RMSEs))])
print(RMSEs[which.min(log(RMSEs))])
res <- apply(preds - y_test, 2, summary)
matplot(res, type = 'l')
abline(h = 0, lty = 2, col = "red")
}
