% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorS3.R
\name{predict.regressor}
\alias{predict.regressor}
\title{Predict using a regressor object}
\usage{
\method{predict}{regressor}(object, newx, ...)
}
\arguments{
\item{object}{A regressor object}

\item{newx}{Matrix or data frame of new observations}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of predicted values
}
\description{
Predict using a regressor object
}
