% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern32.R
\name{predict_matern32}
\alias{predict_matern32}
\title{n <- 100 ; p <- 4}
\usage{
predict_matern32(fit_obj, newx, ci = NULL)
}
\description{
set.seed(456)
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)
}
\details{
lams <- 10^seq(-10, 10, length.out = 50)

fit_obj <- fit_matern32(x = X, y = y, lambda = lams)

df <- data.frame(predict_matern32(fit_obj, newx = X) - y)
colnames(df) <- paste0(round(lams, 2))
summary(df)
boxplot(df[, c(1, 10, 25, 35, 50)], 
main = "distribution of in sample bias", 
xlab = "lambda", ylab = "y_hat - y")
}
