% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressor.R
\name{Regressor}
\alias{Regressor}
\title{`Regressor` class}
\description{
the `Regressor` class contains supervised regression models
}
\section{Super class}{
\code{\link[learningmachine:Base]{learningmachine::Base}} -> \code{Regressor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Regressor-new}{\code{Regressor$new()}}
\item \href{#method-Regressor-fit}{\code{Regressor$fit()}}
\item \href{#method-Regressor-predict}{\code{Regressor$predict()}}
\item \href{#method-Regressor-fit_predict}{\code{Regressor$fit_predict()}}
\item \href{#method-Regressor-clone}{\code{Regressor$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_B"><a href='../../learningmachine/html/Base.html#method-Base-get_B'><code>learningmachine::Base$get_B()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_activ"><a href='../../learningmachine/html/Base.html#method-Base-get_activ'><code>learningmachine::Base$get_activ()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_engine"><a href='../../learningmachine/html/Base.html#method-Base-get_engine'><code>learningmachine::Base$get_engine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_level"><a href='../../learningmachine/html/Base.html#method-Base-get_level'><code>learningmachine::Base$get_level()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_method"><a href='../../learningmachine/html/Base.html#method-Base-get_method'><code>learningmachine::Base$get_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_model"><a href='../../learningmachine/html/Base.html#method-Base-get_model'><code>learningmachine::Base$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_name"><a href='../../learningmachine/html/Base.html#method-Base-get_name'><code>learningmachine::Base$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_nb_hidden"><a href='../../learningmachine/html/Base.html#method-Base-get_nb_hidden'><code>learningmachine::Base$get_nb_hidden()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_nodes_sim"><a href='../../learningmachine/html/Base.html#method-Base-get_nodes_sim'><code>learningmachine::Base$get_nodes_sim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_params"><a href='../../learningmachine/html/Base.html#method-Base-get_params'><code>learningmachine::Base$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_pi_method"><a href='../../learningmachine/html/Base.html#method-Base-get_pi_method'><code>learningmachine::Base$get_pi_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_seed"><a href='../../learningmachine/html/Base.html#method-Base-get_seed'><code>learningmachine::Base$get_seed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="get_type"><a href='../../learningmachine/html/Base.html#method-Base-get_type'><code>learningmachine::Base$get_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_B"><a href='../../learningmachine/html/Base.html#method-Base-set_B'><code>learningmachine::Base$set_B()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_activ"><a href='../../learningmachine/html/Base.html#method-Base-set_activ'><code>learningmachine::Base$set_activ()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_engine"><a href='../../learningmachine/html/Base.html#method-Base-set_engine'><code>learningmachine::Base$set_engine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_level"><a href='../../learningmachine/html/Base.html#method-Base-set_level'><code>learningmachine::Base$set_level()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_method"><a href='../../learningmachine/html/Base.html#method-Base-set_method'><code>learningmachine::Base$set_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_model"><a href='../../learningmachine/html/Base.html#method-Base-set_model'><code>learningmachine::Base$set_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_nb_hidden"><a href='../../learningmachine/html/Base.html#method-Base-set_nb_hidden'><code>learningmachine::Base$set_nb_hidden()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_nodes_sim"><a href='../../learningmachine/html/Base.html#method-Base-set_nodes_sim'><code>learningmachine::Base$set_nodes_sim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_pi_method"><a href='../../learningmachine/html/Base.html#method-Base-set_pi_method'><code>learningmachine::Base$set_pi_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="set_seed"><a href='../../learningmachine/html/Base.html#method-Base-set_seed'><code>learningmachine::Base$set_seed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="learningmachine" data-topic="Base" data-id="summary"><a href='../../learningmachine/html/Base.html#method-Base-summary'><code>learningmachine::Base$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-new}{}}}
\subsection{Method \code{new()}}{
Create a new object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$new(
  name = "Regressor",
  type = "regression",
  model = NULL,
  method = NULL,
  X_train = NULL,
  y_train = NULL,
  pi_method = NULL,
  level = NULL,
  B = NULL,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  seed = 123
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the class}

\item{\code{type}}{type of supervised learning method implemented}

\item{\code{model}}{fitted model}

\item{\code{method}}{supervised learning method in c('lm', 'ranger', 
'extratrees', 'ridge', 'bcn', 'glmnet', 'krr', 'xgboost')}

\item{\code{X_train}}{training set features; do not modify by hand}

\item{\code{y_train}}{training set response; do not modify by hand}

\item{\code{pi_method}}{type of prediction interval in c("splitconformal",
"kdesplitconformal", "bootsplitconformal", "jackknifeplus",
"kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
"surrjackknifeplus")}

\item{\code{level}}{an integer; the level of confidence (default is 95, for 95%)}

\item{\code{B}}{an integer; the number of simulations when \code{level} is not \code{NULL}}

\item{\code{nb_hidden}}{number of nodes in the hidden layer, for construction of a quasi-
randomized network}

\item{\code{nodes_sim}}{type of 'simulations' for hidden nodes, if \code{nb_hidden} > 0; 
takes values in c("sobol", "halton", "unif")}

\item{\code{activ}}{activation function's name for the hidden layer, in the construction 
of a quasi-randomized network; takes values in c("relu", "sigmoid", "tanh", "
leakyrelu", "elu", "linear")}

\item{\code{engine}}{contains fit and predic lower-level methods for the given \code{method}; 
do not modify by hand}

\item{\code{params}}{additional parameters passed to \code{method} when calling \code{fit}
do not modify by hand}

\item{\code{seed}}{an integer; reproducibility seed for methods that include 
randomization}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Regressor` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$fit(
  X,
  y,
  pi_method = c("splitconformal", "kdesplitconformal", "bootsplitconformal",
    "jackknifeplus", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  type_split = c("stratify", "sequential"),
  B = 100,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$predict(X, level = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-fit_predict"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-fit_predict}{}}}
\subsection{Method \code{fit_predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$fit_predict(
  X,
  y,
  pct_train = 0.8,
  score = ifelse(is.factor(y), yes = function(preds, y_test) mean(preds == y_test), no =
    function(preds, y_test) sqrt(mean((preds - y_test)^2))),
  level = NULL,
  pi_method = c("splitconformal", "jackknifeplus", "kdesplitconformal",
    "bootsplitconformal", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  B = 100,
  seed = 123,
  graph = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
