% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierS3.R
\name{classifier}
\alias{classifier}
\title{Create a classifier object for Probabilistic Machine Learning}
\usage{
classifier(
  x,
  y,
  model = c("ranger", "extratrees", "ridge", "bcn", "glmnet", "krr", "xgboost", "svm"),
  pi_method = c("none", "splitconformal", "jackknifeplus", "kdesplitconformal",
    "bootsplitconformal", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  level = 95,
  B = 100,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  seed = 123
)
}
\arguments{
\item{x}{Input matrix or data frame of features}

\item{y}{Vector of target values}

\item{model}{Model to use for classification}

\item{pi_method}{Method to use for conformal prediction}

\item{level}{Confidence level for conformal prediction}

\item{B}{Number of simulations for conformal prediction}

\item{nb_hidden}{Number of nodes in the hidden layer}

\item{nodes_sim}{Type of simulations for hidden nodes}

\item{activ}{Activation function for hidden layer}

\item{engine}{Engine to use for fitting the model}

\item{params}{Additional parameters passed to the model}

\item{seed}{Reproducibility seed for randomization}

\item{...}{Additional arguments passed to Classifier$new()}
}
\value{
A classifier object of class "classifier"
}
\description{
Create a classifier object for Probabilistic Machine Learning
}
\examples{

X <- iris[, -5]
y <- iris$Species
X_train <- X[1:100, ]
y_train <- y[1:100]
X_test <- X[101:150, ]
y_test <- y[101:150]
clf <- classifier(X_train, y_train, pi_method = "kdesplitconformal", level = 95)
predict(clf, newx = X_test)
print(mean(y_test == predict(clf, newx = X_test)))

}
