% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorS3.R
\name{regressor}
\alias{regressor}
\title{Create a regressor object for Probabilistic Machine Learning}
\usage{
regressor(
  x,
  y,
  model = c("ranger", "extratrees", "ridge", "bcn", "glmnet", "krr", "xgboost", "svm"),
  pi_method = c("none", "splitconformal", "jackknifeplus", "kdesplitconformal",
    "bootsplitconformal", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  level = 95,
  B = 100,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  type_split = c("stratify", "sequential"),
  seed = 123
)
}
\arguments{
\item{x}{Input matrix or data frame of features}

\item{y}{Vector of target values}

\item{model}{Model to use for regression}

\item{pi_method}{Method to use for conformal prediction}

\item{level}{Confidence level for conformal prediction}

\item{B}{Number of simulations for conformal prediction}

\item{nb_hidden}{Number of nodes in the hidden layer}

\item{nodes_sim}{Type of simulations for hidden nodes}

\item{activ}{Activation function for hidden layer}

\item{engine}{Engine to use for fitting the model}

\item{params}{Additional parameters passed to the model}

\item{type_split}{Type of data splitting for split conformal prediction:
"stratify" (for classical supervised learning) "sequential" (when
the data sequential ordering matters)}

\item{seed}{Reproducibility seed for randomization}

\item{...}{Additional arguments passed to Regressor$new()}
}
\value{
A regressor object of class "regressor"
}
\description{
Create a regressor object for Probabilistic Machine Learning
}
\examples{

X <- mtcars[, -1]
y <- mtcars$mpg
X_train <- X[1:25, ]
y_train <- y[1:25]
X_test <- X[26:32, ]
y_test <- y[26:32]
reg <- regressor(X_train, y_train, pi_method = "splitconformal", level = 95)

print(sqrt(mean((y_test - predict(reg, newx = X_test))^2)))

}
